# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Debt Relief & Financial Recovery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialAnalysis } from './financialAnalysis';
import type { Loan } from './loan';
import type { LoansByStatus } from './loansByStatus';
import type { MonthlyTrendItem } from './monthlyTrendItem';

export interface DashboardSummary {
  totalDebt: number;
  /** @nullable */
  monthlyIncome?: number | null;
  /** @nullable */
  monthlyExpenses?: number | null;
  /** @nullable */
  financialScore?: number | null;
  /** @nullable */
  debtRatio?: number | null;
  /** @nullable */
  settlementProbability?: number | null;
  loanCount: number;
  totalMonthlyPayment: number;
  recentLoans: Loan[];
  recentAnalyses: FinancialAnalysis[];
  loansByStatus: LoansByStatus;
  monthlyTrend: MonthlyTrendItem[];
}