# lib/api-zod/src/generated/types/financialAnalysis.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Debt Relief & Financial Recovery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FinancialAnalysis {
  id: number;
  userId: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  monthlySavings: number;
  financialScore: number;
  debtRatio: number;
  riskLevel: string;
  recommendations: string[];
  disposableIncome: number;
  totalMonthlyDebt: number;
  createdAt: string;
}